New tests made passing requirements

+++ back-end/src/database/migrations/20220222184017-create-sales.js
@@ -26,7 +26,7 @@ module.exports = {
       },
       total_price: {
         allowNull: false,
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL(9,2),
         field: 'total_price'
       },
       delivery_address: { allowNull: false, type: Sequelize.STRING, field: 'delivery_address' },
